performance finish

+++ app/src/main/java/com/dr/jjsembako/core/data/remote/response/performance/GetFetchOmzetResponse.kt
@@ -23,5 +23,5 @@ data class OmzetData(
 	val month: Int,
 
 	@field:SerializedName("omzet")
-	val omzet: Int
+	val omzet: Long
 )
